requirements.txt
aiofiles==0.4.0
aiomysql==0.0.20
appdirs==1.4.3
arrow==0.15.4
asgiref==3.2.3
asn1crypto==1.2.0
aspy.yaml==1.3.0
attrs==19.3.0
Automat==0.8.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
binaryornot==0.4.4
bleach==3.1.0
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.0
cfgv==2.0.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
conda==4.8.0
conda-package-handling==1.6.0
constantly==15.1.0
cookiecutter==1.6.0
cryptography==2.8
cssselect==1.1.0
ddt==1.2.2
decorator==4.4.1
Django==3.0.2
django-debug-toolbar==2.2
django-environ==0.4.5
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
docopt==0.6.2
dpkt==1.9.2
environ==1.0
environs==7.1.0
fake-useragent==0.1.11
filelock==3.0.12
future==0.18.2
geoip2==3.0.0
hyperlink==19.0.0
identify==1.4.9
idna==2.6
importlib-metadata==1.3.0
incremental==17.5.0
Jinja2==2.10.3
jinja2-time==0.2.0
lxml==4.4.2
MarkupSafe==1.1.1
marshmallow==3.3.0
maxminddb==1.5.2
menuinst==1.4.16
more-itertools==8.0.2
multi-key-dict==2.0.3
mysqlclient==1.4.6
nodeenv==1.3.3
packaging==19.2
paramiko==2.7.1
parse==1.14.0
pbr==5.4.4
peewee==2.10.2
peewee-async==0.5.12
Pillow==7.0.0
pipreqs==0.4.10
pluggy==0.13.1
poyo==0.5.0
pre-commit==1.20.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycosat==0.6.3
pycparser==2.19
pyee==6.0.0
PyHamcrest==1.9.0
PyJWT==1.7.1
PyMySQL==0.9.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyppeteer==0.0.25
pyquery==1.4.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-jenkins==1.6.0
python-memcached==1.59
pytz==2019.3
pywin32==223
PyYAML==5.2
qiniu==7.2.6
redis==3.3.11
requests==2.22.0
requests-file==1.4.3
requests-html==0.10.0
retry==0.9.2
robotframework==3.1.2
robotframework-seleniumlibrary==4.1.0
ruamel-yaml==0.15.46
scapy==2.4.3
selenium==3.8.1
service-identity==18.1.0
shortuuid==0.5.0
simplejson==3.17.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
sqlparse==0.3.0
testrail==0.3.12
toml==0.10.0
Top==1.0.0
torext==0.10.0
tornado==6.0.3
tox==3.14.3
tqdm==4.36.1
translate==3.5.0
Twisted==19.10.0
urllib3==1.25.7
virtualenv==16.7.9
w3lib==1.21.0
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.0
whichcraft==0.6.1
win-inet-pton==1.1.0
wincertstore==0.2
WTForms==2.2.1
WTForms-JSON==0.3.3
wtforms-tornado==0.0.2
xlrd==1.2.0
XlsxWriter==1.0.5
yarg==0.1.9
zipp==0.6.0
zope.interface==4.7.1
